allowances: tests: start fixing tests, things are still missing
